update ft_uint_to_hex & ft_uint_to_octal to handle uintmax_t values - update header & convert arg modifiers to handle all 'l' conversions

+++ includes/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: jjaniec <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/12/11 13:41:09 by jjaniec           #+#    #+#             */
-/*   Updated: 2018/01/21 16:13:46 by jjaniec          ###   ########.fr       */
+/*   Updated: 2018/01/24 13:07:43 by jjaniec          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -56,9 +56,9 @@ char	*ft_char_to_str(char c);
 
 char	*ft_uint_to_str(unsigned int x);
 
-char	*ft_uint_to_hex(unsigned int n);
+char	*ft_uint_to_hex(uintmax_t n);
 
-char	*ft_uint_to_hex_caps(unsigned int n);
+char	*ft_uint_to_hex_caps(uintmax_t n);
 
 void	ft_putwchar(wchar_t c);
 
@@ -74,7 +74,7 @@ char	*ft_wchar_tptr_to_str(wchar_t *ws);
 
 char	*ft_wchar_t_to_str(wchar_t c);
 
-char    *ft_uint_to_octal(const unsigned int n);
+char    *ft_uint_to_octal(const uintmax_t n);
 
 char    *ft_long_int_to_uoctal(const long int n);
